Moving heroes array generator to top level component.
